requirements.txt
loguru==0.6.0
kafka-python==2.0.2
pydantic==1.10.2
python-dotenv==0.21.0